requirements.txt
bioframe
matplotlib
numpy
pandas
pybbi
pyjaspar
scikit-learn
tqdm